Chore: Move Prometheus datasorce tests from specs folder and merge duplicated test files
