+++ public/app/plugins/datasource/prometheus/module.test.ts
@@ -1,4 +1,4 @@
-import { plugin as PrometheusDatasourcePlugin } from '../module';
+import { plugin as PrometheusDatasourcePlugin } from './module';
 
 describe('module', () => {
   it('should have metrics query field in panels and Explore', () => {
